pyproject.toml
[project]
name = "ticker-commentator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "huggingface-hub",
    "numpy",
    "pandas",
    "plotly",
    "snac",
    "torch>=2.10",
    "triton-rocm>=3.6.0; sys_platform == 'linux'",
    "streamlit",
    "yfinance",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "llama-cpp-python>=0.3.16",
    "python-dotenv",
    "pytest>=8.4.0",
]

[tool.uv.sources]
torch = { index = "pytorch-rocm" }
triton-rocm = { index = "pytorch-rocm" }
torchaudio = { index = "pytorch-rocm" }
torchvision = { index = "pytorch-rocm" }


[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/rocm7.1"
explicit = true
